requirements.txt
ruamel.yaml>=0.18.0
rich>=13.0.0
